+++ khoros/core.py
@@ -1713,6 +1713,20 @@ class Khoros(object):
             """
             return structures_module.boards.get_message_count(self.khoros_object, board_id)
 
+        def get_all_messages(self, board_id, fields=None):
+            """This function retrieves data for all messages within a given board.
+
+            .. versionadded:: 5.3.0
+
+            :param board_id: The ID of the board to query
+            :type board_id: str
+            :param fields: Specific fields to query if not all fields are needed (comma-separated string or iterable)
+            :type fields: str, tuple, list, set, None
+            :returns: A list containing a dictionary of data for each message within the board
+            :raises: :py:exc:`khoros.errors.exceptions.GETRequestError`
+            """
+            return structures_module.boards.get_all_messages(self.khoros_object, board_id, fields)
+
     class BulkData(object):
         """This class includes methods for interacting with the Bulk Data API.